feat: leitura e escrita modbus melhorados, falta criação de banco de dados MongoDb
