+++ server/update/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "update",
+  "version": "0.0.1",
+  "description": "reading the current status of the PLC",
+  "main": "leituraCLP.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
